*** empty log message ***

+++ coders/pcd.c
@@ -280,7 +280,7 @@ static unsigned int DecodeImage(Image *image,unsigned char *luma,
     r=pcd_table[plane];
     for (i=0; ((i < (int) length) && ((sum & r->mask) != r->sequence)); i++)
       r++;
-    if (((q-luma) > (size_t) (image->columns*image->rows)) ||
+    if (((size_t) (q-luma) > (size_t) (image->columns*image->rows)) ||
         (r == (PCDTable *) NULL))
       {
         ThrowException(&image->exception,CorruptImageWarning,
